requirements.txt
pillow
newspaper3k
urllib3
streamlit
beautifulsoup4
protobuf==3.20.0
jinja2>=3.0
langchain_community
langchain_google_genai
langchain_core
hub
beautifulsoup4
urllib
newspaper
io
nltk
fitz
PyPDF2
googleapiclient
spacy
streamlit_option_menu
